requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
ansicon==1.89.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argostranslate==1.9.1
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
ctranslate2==3.20.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
faster-whisper==0.9.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
fqdn==1.5.1
fsspec==2023.10.0
future==0.18.3
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.28.1
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keyboard==0.13.5
lazy_loader==0.2
llvmlite==0.41.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
onnxruntime==1.16.2
openai-whisper==20231106
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.0
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
pytweening==1.0.7
pyuseragents==1.0.5
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
regex==2023.10.3
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
sacremoses==0.0.53
safeIO==1.2
safetensors==0.4.0
scikit-image==0.20.0
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.2
stanza==1.1.1
sympy==1.12
terminado==0.17.1
tifffile==2023.4.12
tiktoken==0.5.1
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.14.1
torch==2.1.0
tornado==6.3.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.1
translatepy==2.3
typing_extensions==4.8.0
uri-template==1.2.0
urllib3==2.0.2
Wave==0.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
y-py==0.5.9
ypy-websocket==0.8.2
